+++ iris_camera_toolkit/sobol_image_sampler.h
@@ -28,7 +28,7 @@ extern "C" {
 
 ISTATUS
 SobolImageSamplerAllocate(
-    _In_ uint16_t samples_per_pixel,
+    _In_ uint32_t samples_per_pixel,
     _Out_ PIMAGE_SAMPLER *image_sampler
     );
